allow passing of a custom client
The standard way this works is that you tell the worker to start
working, and it creates its own client for doing so. The redis
connection is created automatically from a REDIS_URL environment
variable. Unfortunately that doesn't work at all when we want to use
sentinel because the redis client can't create itself from a sentinel
URL. This adds a new way for creating a worker where we pass in our own
client, giving us the ability to configure it however we want to.