Fix failure formatter when pwd is a symlink.
`Dir.pwd` uses `getcwd` under the hood, which resolves the symlink if
pwd is such. This is commonly the case for production Rails apps, where
a `current` symlink points to a release directory. In contrast,
`ENV['PWD']` will be the pwd at the time the app was launched, and does
not attempt to resolve any symlinks.

Aside from being susceptible to having the wrong value if a job leaked a
pwd change, this also fixes a failure in the case the original release
directory is swept before a long running job throws an error.